Add alternative, key-centric API
This new API returns Promises instead of accepting callbacks. It takes
some inspiration from the Web Crypto API in that it accepts any
BufferSource and outputs ArrayBuffers.

Additionally, public and private keys can only be used through specific
classes, which can be instantiated by importing existing key material or
by generating new key pairs. This makes the API difficult to misuse, and
the native addon can keep key material out of the JavaScript heap as
much as possible.

Because this API does not rely on any Node.js-specific types or
concepts, it might be possible for it to support other runtimes (at
least using the WebAssembly backend instead of the native addon).